requirements.txt
alembic==0.9.7
antiorm==1.2.1
attrs==17.4.0
click==6.7
Flask==0.12.1
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.7
MarkupSafe==1.0
pluggy==0.6.0
py==1.5.2
PyMySQL==0.8.0
python-dateutil==2.6.1
python-editor==1.0.3
python-memcached==1.59
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.12.1
mysql-connector-python==8.0.5
